+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * rewrite maintenance script ([#243 by @agross](https://github.com/buanet/ioBroker.docker/pull/243))
 * enhance container shutdown on SIGTERM ([as requested with #264 by @buzz0r](https://github.com/buanet/ioBroker.docker/pull/264))
 * enhance startup script logging
+* enhance logging for avahi & zwave install scripts
 
 ### v6.1.0 (2022-03-01)
 * v6.1.0-beta.2 (2022-02-11)